refactor(gsd): extract auto-worktree teardown module
